user and system prompt
